+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * The annotation processor now correctly reports an error if trying to reference interfaces in model classes (#2808).
 * Added null check to `addChangeListener` and `removeChangeListener` in `Realm` and `DynamicRealm` (#2772).
 * Calling `RealmObjectSchema.addPrimaryKey()` adds an index to the primary key field, and calling `RealmObjectSchema.removePrimaryKey()` removes the index from the field (#2832).
+* Log files are not deleted when calling `Realm.deleteRealm()` (#2834).
 
 ### Enhancements